Move auth.entity.UserEntity -> user.entity.UserEntity, auth.entity.Role -> user.entity.Role
