+++ src/main/java/site/hirecruit/hr/domain/auth/dto/AuthUserInfo.kt
@@ -5,7 +5,7 @@ import org.springframework.context.annotation.Scope
 import org.springframework.context.annotation.ScopedProxyMode
 import org.springframework.stereotype.Component
 import org.springframework.web.context.WebApplicationContext
-import site.hirecruit.hr.domain.auth.entity.Role
+import site.hirecruit.hr.domain.user.entity.Role
 
 /**
  * 인증/인가 시 유저의 정보를 담고 있는 객체
