+++ src/main/java/site/hirecruit/hr/domain/auth/repository/UserRepository.kt
@@ -1,7 +1,7 @@
 package site.hirecruit.hr.domain.auth.repository
 
 import org.springframework.data.jpa.repository.JpaRepository
-import site.hirecruit.hr.domain.auth.entity.UserEntity
+import site.hirecruit.hr.domain.user.entity.UserEntity
 
 interface UserRepository : JpaRepository<UserEntity, Long>, UserCustomRepository {
     fun existsByGithubId(githubId: Long) : Boolean
