+++ src/main/java/site/hirecruit/hr/domain/auth/service/impl/UserRegistrationRollbackServiceImpl.kt
@@ -5,7 +5,7 @@ import org.springframework.transaction.annotation.Propagation
 import org.springframework.transaction.annotation.Transactional
 import site.hirecruit.hr.domain.auth.dto.AuthUserInfo
 import site.hirecruit.hr.domain.auth.dto.OAuthAttributes
-import site.hirecruit.hr.domain.auth.entity.Role
+import site.hirecruit.hr.domain.user.entity.Role
 import site.hirecruit.hr.domain.auth.repository.UserRepository
 import site.hirecruit.hr.domain.auth.service.TempUserRegistrationService
 import site.hirecruit.hr.domain.auth.service.UserRegistrationRollbackService
