+++ src/main/java/site/hirecruit/hr/domain/mentor/aop/MentorServiceAspect.kt
@@ -5,7 +5,7 @@ import org.aspectj.lang.annotation.Aspect
 import org.aspectj.lang.annotation.Pointcut
 import org.springframework.stereotype.Component
 import site.hirecruit.hr.domain.auth.dto.AuthUserInfo
-import site.hirecruit.hr.domain.auth.entity.Role
+import site.hirecruit.hr.domain.user.entity.Role
 import site.hirecruit.hr.domain.auth.service.SecurityContextAccessService
 import site.hirecruit.hr.global.data.SessionAttribute
 import javax.servlet.http.HttpSession
