+++ src/main/java/site/hirecruit/hr/domain/worker/repository/WorkerCustomRepositoryImpl.kt
@@ -3,7 +3,6 @@ package site.hirecruit.hr.domain.worker.repository
 import com.querydsl.core.types.dsl.Expressions
 import com.querydsl.jpa.impl.JPAQueryFactory
 import site.hirecruit.hr.domain.auth.entity.QUserEntity.userEntity
-import site.hirecruit.hr.domain.auth.entity.Role
 import site.hirecruit.hr.domain.company.dto.QCompanyDto_Info
 import site.hirecruit.hr.domain.company.entity.QCompanyEntity.companyEntity
 import site.hirecruit.hr.domain.worker.dto.QWorkerDto_Info
