+++ src/main/java/site/hirecruit/hr/global/dummy/SetDummyData.kt
@@ -3,8 +3,8 @@ package site.hirecruit.hr.global.dummy
 import net.bytebuddy.utility.RandomString
 import org.springframework.context.annotation.Configuration
 import org.springframework.context.annotation.Profile
-import site.hirecruit.hr.domain.auth.entity.Role
-import site.hirecruit.hr.domain.auth.entity.UserEntity
+import site.hirecruit.hr.domain.user.entity.Role
+import site.hirecruit.hr.domain.user.entity.UserEntity
 import site.hirecruit.hr.domain.auth.repository.UserRepository
 import site.hirecruit.hr.domain.company.entity.CompanyEntity
 import site.hirecruit.hr.domain.company.repository.CompanyRepository
