+++ src/test/java/site/hirecruit/hr/domain/auth/aop/UserSessionInfoUpdateAspectTest.kt
@@ -10,7 +10,7 @@ import org.springframework.aop.aspectj.annotation.AspectJProxyFactory
 import org.springframework.mock.web.MockHttpSession
 import site.hirecruit.hr.domain.auth.dto.AuthUserInfo
 import site.hirecruit.hr.domain.auth.dto.OAuthAttributes
-import site.hirecruit.hr.domain.auth.entity.Role
+import site.hirecruit.hr.domain.user.entity.Role
 import site.hirecruit.hr.domain.auth.service.UserAuthService
 import site.hirecruit.hr.domain.auth.service.UserRegistrationRollbackService
 import site.hirecruit.hr.global.data.SessionAttribute
